refactor: rename Permission -> PermissionHistory | move mixin to same file
